[Example] Split Stream-K kernel into dedicated example and address review feedback
